+++ packages/morphic-show-interpreters/src/interpreters.ts
@@ -10,8 +10,10 @@ import { showRefinedInterpreter } from './model/refined'
 import { showSetInterpreter } from './model/set'
 import { showStrMapInterpreter } from './model/str-map'
 import { showTaggedUnionInterpreter } from './model/tagged-unions'
+import { showUnionInterpreter } from './model/unions'
 import { showUnknownInterpreter } from './model/unknown'
 
+export {} from './model/unions'
 export {} from './model/intersections'
 export {} from './model/newtype'
 export {} from './model/object'
@@ -30,6 +32,7 @@ export * from './hkt'
  */
 const allModelShow = <Env extends AnyEnv>() =>
   merge(
+    showUnionInterpreter<Env>(),
     showRefinedInterpreter<Env>(),
     showNewtypeInterpreter<Env>(),
     showUnknownInterpreter<Env>(),